2.2.3: make the fee uint128 to accept values greater than 18 xDAI

+++ contracts/ProxyEvent.sol
@@ -6,7 +6,7 @@ contract ProxyEvent is Event {
 
     mapping(address => address) public proxy;
 
-    constructor(uint64 _fee, uint32 _end) Event(_fee, _end) public {}
+    constructor(uint128 _fee, uint32 _end) Event(_fee, _end) public {}
 
     function registerFor (address payable _attendee) external payable {
         register(_attendee, msg.value);
